fix (pages/about) Moves styles for page to another separate folder to prevent exposing of non-page files on build
